lang,c: fix management of the type stack
buildCurrentType should obviously pop the top type on the stack. Do that
and fix misbehaviours related to places where the top type was reused.

at end of statements, the type stack should be empty. Because of some
issues (mainly the fact that we don't handle function definitions), this
is not the case. Force it.