core,test: fix merging of compounds
It is possible for a Compound to be bigger than the sum of its fields (C
compilers may round struct sizes for instance). Fix the merging algorithm by
explicitely updating the Compound size just after the merge.

This commit also introduces a test where the C-lang test repository is merged
into an empty repository and into itself, and checks the result. It gives a very
broad testing of the merging algorithm as a lot of different types are defined
in this file (even thought the failures would not be very specific).