gracefully handle template markers in name manipulations
First, types inserted in between the template markers '<' and '>'
where wrongfully taken as part of the type and namespace
names. This commit fixes it, so that

  /Bla/Blo</Template>

is split into
  /Bla/
  /Blo</Template>

Moreover, in the Ruby library, #basename and #full_name now
replace the namespace markers in the template arguments as well, i.e.
one gets
  ::Bla::Blo<::Template>

for the previous example.