ruby: do not rely on Type objects to destroy memory zones
We have to call Typelib::destroy() on memory zones in order
to de-initialize containers. Right now, that was done by storing
the Type object of each memory zone and call Typelib::destroy(value).

The problem is that, at exit, the Ruby GC calls finalizers in an
un-ordered way. Therefore, the type registry may get deleted *before*
the memory zones, leading to a crash.

This commit now pre-computes the memory layouts at memory initialization.
In order to save time and space, the layout of one specific type is
computed only once.