test: fix longstanding issues with float rounding in to_csv tests
to_csv unit tests where failing because floats where rounded
un-exactly and therefore the string comparison where failing.

This test fixes it by doing string-to-float conversion and
checking the rounded value directly. One test disappeared
from the C++ test suite because it was too cumbersome to do
there, but the corresponding code is also tested in the Ruby suite.