idl: fix namespace management in IDL exporter
In a3b32ea66f59ef3790429597150b5edfd412bacd, we made sure that the exporter
would not generate empty modules (that are not valid in IDL). For that, we
delayed the call to adaptNamespace() to the point where we know we have some
definitions to put in.

But the IDLExportVisitor actually uses the m_namespace option to generate the
right type references (i.e. adding explicit namespace in front of the types when
needed). That was broken.

This workaround temporarily changes the m_namespace attribute while the type
definitions are generated in order to make things up. This is a bit ugly also.