ruby: define the NotFound exception for #get and #build
Until now, #get was returning nil if the object did not exist,
and #build was returning TypeError.

This commit defines the NotFound exception, and makes both methods
raise it in case the type cannot be fetched or build. This makes
the error handling on #get less tricky.

The goal, in the end is to unify all error cases "did not find something"
under the NotFound exception.