fix alias copy in minimal()
Before copying an alias, minimal() is checking if the target name is already in
use (and also checks that the concrete type is the same). Fine.

The issue was that the test was done on the source registry, not on the target
one. So, there was always a definition and it always matched, and therefore the
aliases were never copied.

This fixes the issue