tlb: properly save and reload container size
Containers a plugged-in, which means that their size depend
on the actual architecture on which the data is generated.

This is fine since the marshalled data is architecture
independent, but -- still -- it means that a registry should
be updated to reflect the size change of containers.

To allow this, save and reload container size in .tlb, and fix
the container size in the registry. The user will have to call
resize() if he wants to use those types.