we now should only use absolute targets in configure_file
CMake 2.8.0 changes its behaviour w.r.t. how the target file is interpreted in
configure_file. The old behaviour was that relative path are relative to the
root build dir. The new one that they are relative to the current binary dir.

So, just use absolute paths for now on.