build: add some missing links between shared libraries
The typelib plugins need to be linked to the main typelib library. This is
actually not required, as they are loaded as plugins, but it does not harm to do
link and allows better diagnostics of problems with undefined symbols.

Also link the ruby extension with libruby.so, for the same reasons.