idl: workaround a bug in the omniORB IDL compiler
In some cases, when modules are reopened, the IDL compiler "forgets" to generate
typecode information for some typedefs.

The workaround is to generate typedefs at the end. Indeed, since the typedef
code generation "flattens" the types (i.e. typedefs are not used in other IDL
types), we can postpone the typdef code generation at the end of the IDL
generation, thus limiting closing and reopening modules.