fix merging of Container
When we were merging Container types, we were only creating a new
container, not caring about its size. Therefore, the container
was created with the natural platform size -- which can be different
from the actual size in the source registry.

Fix that by explicitely setting the size of the new container type.