ruby: add support for custom conversions and type specializations
This commit adds Typelib.convert_from_ruby, Typelib.specialize and
Typelib.specialize_model

The first one declares that a given ruby type can be converted into
the corresponding Typelib type with the given block.

The two others declare methods and singleton methods (respectively)
that should be defined on the Class instance that represents a
given type.