fix some issues with containers of containers
Not all resize operations were properly init'ing and destroy'ing
old and new elements, which was leading to crashes and memory leaks
when manipulating containers.

Moreover, we stop assuming that memcpy(&vector_dst, &vector_src) is
valid if we do not destroy vector_src. Instead, if a container element
is removed, we properly copy the other elements to erase it and destroy
the last element of the container.