remove code related to ruby-marshalling
Typelib values should not be marshalled "as-is", and should
definitely not be marshalled through Ruby as we would have to
marshal the registry with them

This code was an attempt at transmitting Typelib values through
DRb, which never worked. So remove it and resurrect it if it is
actually needed some time later.