ruby: add support for plugins
The TYPELIB_PLUGIN_PATH environment variable is now searched for
ruby files by Typelib. These files are required at the first Registry
initialization, and can define extensions to the typelib types and/or
convertions from Ruby to Typelib (using specialize_model, specialize
and convert_from_ruby)

Moreover, a Typelib::DynamicWrapper method can be used to bind a
Typelib type with its Ruby equivalent, allowing to directly set fields
even AND manipulate the type as a ruby object.

For instance, with

  Typelib.specialize_model '/base/Time' do
    def to_ruby(value)
      Typelib::DynamicWrapper(
         Time.at(value.seconds, value.microseconds),
         value)
    end
  end

The value of "my_structure.time" will mostly behave like an instance
of Time (ruby object), but one will be able to do

  my_structure.time.seconds = new_value

as if my_structure.time was still the Typelib value for /base/Time