core: rework the main export algorithm
Until now, the export algorithm was picking types that could be
exported one by one (i.e. types whose dependencies have been
exported already). This had the drawback of not minimizing the
number of namespace changes that were required.

Change the algorithm: gather all the "free" types -- types whose
dependencies are exported already -- in a TypeMap that is sorted
on the basis of the namespace + type name and update that typemap
at each export.

This is very rough, but works well in most cases.

It works around some bugs of omniidl when modules are closed and
reopened.