implement auto-adding aliases for indirect types
Whenever we create an indirect type, a set of implicit aliases
exist -- built from the aliases of the pointed-to type.

However, these aliases are not explicitely created. What it means
is that everything is fine as long as the referred-to type stays
a concrete type. However, if at some point it changes from a
concrete to a typedef, then the indirect type cannot be created
anymore.

As an example, let's consider /std/vector</int>

As long as /int remains a concrete type, /std/vector</int> exists.

If (something that just happened), /int is changed to an alias, then
Container(/std/vector, /int) is *not* named /std/vector</int> anymore
as registry.get('/int') resolves to the concrete type. Ergo, anything
that was referring to /std/vector</int> by name stops working.

This commit fixes the problem by having a generic mechanism through
which types are asked to create/update aliases for themselves whenever
one of their direct dependencies is aliased.