core: improve the handling of the persisten flag a bit
add() and alias() now both have a version where the persistent
flag is specified, and this flag is used to mark autogenerated
aliases of arrays, pointers an containers as non-persistent.

However, the persistent flag of concrete types is still not kept
across merges: if a type A is non-persistent in the source registry,
it will still be marked as persistent in the final merged registry.
It works fine for typedefs though