lang,cxx: registrer /std/string only at import time
Having it always registered causes many problems (actually, no
types should be registered by default, but that patch will be
a lot bigger).

For now, fix the immediate issue of /std/string by moving the
registration of standard C++ types in a separate function which
is called at loading time.