idl: rework the IDL name generation
The end goal of this commit is to have the generated IDL use
typedefs for sequences used in structs instead of using
sequence<> directly. The reason is that the C++ type used for
sequence<>, when used directly, is not uniform across CORBA
implementations while the typedef-related type is.

To implement this, a good cleanup of the IDL exporter has been
in order. In particular, the name generation is now a lot more
straightforward than before -- where multiple methods where used
in a completely non-uniform manner.

This will however remove the workarounds set in place for the
omniorb IDL compiler. Another commit that puts it back in place
will come.