idl: non-hacky way to tell the exporter "this is a builtin type"
If the local identifier namespace  is set to empty, this gets
interpreted as "this is a builtin type, don't add a namespace
specification". However, the emission of sequence<> is modified
so that the sequence is still emitted in orogen::Corba, even
for these types.

This fixes a problem with opaques_as_any, as the any's were referred
to as orogen::Corba::any