fix setting fields in structures that are std::string
The issue is that std::string is implemented as a container
in typelib, and therefore CompoundType[]= uses #set_values
to set the field.

However, it led to do
  field_value.insert(string)

which (1) wrongly accepts its first argument and (2) only
inserts one character as /std/string is a container of bytes.

This commits specializes #set_values for /std/string to do
the right thing.