introduce a workaround for GCCXML handling of templates
Gccxml does not handle templates. So, in order to see template
instances, one needs to *really instanciate* the template.

By *really instanciate*, I mean that typedefs are not enough. You
need to define a field in a structure or a global variable with
said template value.

In order for it to not show up too much in the resulting registries,
all types whose name contains gccxml_workaround are filtered out.