gccxml: fix resolving contained types
GCCXML is a bit inconsistent when it comes to how types are
described. Classes have a "demangle" field in which the full
type name is included. Enums and typedefs don't.

The net result is that we must make sure that, for those types,
we search on both basename and context (i.e. namespace). This
commit introduces a generic method to resolve the namespace,
and fixes resolution of element names for containers (which showed
that particular bug)