ruby: major cleanup of the type convertion infrastructure
The abilities of the Ruby bindings to convert between Typelib and
Ruby types have grown over time. It has also never been cleaned up,
leaving a mess of different code paths that are supposed to deal
with the same problem.

Rationalize. The convertions are now handled by Typelib.to_ruby
and Typelib.from_ruby, with Type#to_ruby a shortcut to
Typelib.to_ruby.

Moreover, Typelib will now handle the ruby<->C++ convertion as if
the convertions were defining a Ruby mapping from the C++ typespace.
In practice, it means that the type

  struct A
  {
     B field;
  };

will be mapped using the defined convertion. I.e. if there is a convertion
defined for 'B', then the value returned for values of A will *not* do
an on-the-fly convertion, but will instead behave as if

  class A
    attr_reader :field
  end

was used, with @field initialized by the convertion of the value of B

This makes the use of the custom convertions completely natural in the
Ruby scripts.