core: define splitTypename, fix getTypename and getNamespace for template-as-namespaces
Both methods would fail if a template was present in the path to a type
(as opposed as being the type itself). The reason was that both get*
functions were not parsing the value properly.

Introduce splitTypename which splits a type name in its elements,
and use it to fix getTypename/getNamespace.