ruby: fix handling of recursive type convertion
there is now "partial" type convertion (kind of). I.e. the
Typelib type objects get extended to only copy the fields that
need to be copied.

However, once one type behaves like this, it is important that the
types that contain it understand that it shoul be handled as a converted
type (by opposition to a normal type).

So, instead of using Type.convertion_to_ruby as a trigger for recursive
type convertion, set a #contains_converted_types? flag explicitely and
use it.