ruby: add a way to make sure that all converted fields get updated from the underlying typelib value
The convert-to-ruby mechanism works by creating a converted value and caching
it for subsequent calls to the field/...

For instance, with

  struct Sample
  {
     base::Time time;
  };

then sample.time (in Ruby) will be converted once to a Time instance,
and that instance will always get returned (no convertion takes place
anymore).

However, if one wants to update that typelib value from C/C++, then
the ruby side needs to know that it should update the converted-to-ruby
values as well. One should therefore call #invalidate_changes_from_converted_types
to make sure that the update occurs.