ruby: fix #(invalidate,apply)_changes_from_converted_types for compound-in-compound-with-converted-field
The methods was not call recursively on fields that are themselves
Typelib compounds and contain converted types. In effect, it meant
that invalidate_ and apply_ would not work on types that look
(for instance) like

  struct B
  {
     base::Time time;
  }
  struct A
  {
     B field;
  };

where base::Time is being converted to Ruby time. The symptoms
were that, if the raw data of A.field was modified, that was
not reflected in the converted type even if
invalidate_changes_from_converted_types was called.

Idem for modified converted fields where modifications would not
get propagated to the raw data even if apply_changes_... is called.

Signed-off-by: Sylvain Joyeux <sylvain.joyeux@dfki.de>