ruby: fix handling of false as a valid result of the convertion to ruby
The issue was that the convertion handling code was using


  if !(v = instance_variable_get(@field_name))

to check whether a particular field had already been converted. This
is wrong if we want 'false' to be a possible value.

Instead, check for nil-ness

Ideally, we would use instance_variable_defined?, but there are
no ways I know of to un-define an instance variable (something
we need to do in #invalidate_changes_from_converted_types