ruby: properly invalidate values of modified containers
This commit implements a freeze/invalidate functionality that allows
to limit access to certain Typelib values to resp. read-only or
no-access.

The latter is used to invalidate "child" values when an operation
on the parent demands it. The canonical example is the resizing of
std::vector, which might invalidate any child value.

For now, a side effect of this is that the objects managing the child
elements are all kept. If it becomes a problem in practice, we will
have to use weak references to ensure that they can be GCed