ruby: fix handling of invalidation
The main issue here was that ArrayType and ContainerTYpe would
not always return the same wrapper, i.e.:

  container[10].object_id != container[10].object_id

Since @__typelib_children was defined on the wrapper instance,
invalidation would not work under some circumstances. Note that
it was mostly working since ContainerType *does* return the same
wrapper for the same field.

This commit makes ArrayType and ContainerType save non-plain fields,
and uses the hashes created by that functionality to invalidate the
children (thus removing the @__typelib_children hash).

Moreover, it normalizes the element access interfaces around two
methods: raw_get and raw_set, for all the ContainerType, ArrayType
and ContainerType, and makes sure that the access and caching
logic is there and only there. The test suite got updated to reflect
that design constraint.