ruby: fix char-to/from-string convertions when /char is not available
One should resolve all types to the normalized (u)int_X types. In
this case, /char is not necessarily available at convertion time.

However, since we verify that we convert the "right" type to string
in typelib.rb first, we can bypass the typecheck in the C extension.
Do so.