ruby: fix memory handling w.r.t. the lazy-sweep GC in ruby 1.9.3
Ruby 1.9.3 does not sweep all values in one shot anymore, and -- as before --
the sweeping order in undeterministic.

This means that it is now possible that a root value gets GCed, a child value is
not swept (even though it is not alive anymore) and the VM returns to client
code. Then, the deallocated memory is reused and typelib memory management
thinks that we have a case for double-wrappers.

This commit implements a refcounting mechanism to handle the root memory zones,
making sure that we never deallocate a root memory zone until all the other
zones referring to it are properly deallocated.