ruby: handle the fact that some elements in #handle_container_invalidation might be deleted
When applied on a container of a complex type, which contains other
containers, it is possible that the toplevel container deleted
some elements. These elements should therefore be ignored.

This is done in two places. #handle_container_invalidation ignores
already-invalidated elements while Typelib.handle_invalidation makes
sure that we process the (in)validation of the toplevel elements
first