ruby: fix type resolution in one nested-type use case (closes #321)
This is a fix for a bug triggered if:
 - one nested type is not defined in a direct header
   (it has to be included by the file that is imported
   in orogen)
 - it is used in the struct that defines it and
 - it is used before the containing type in one of the
   structs that is in the header directly imported in orogen

The issue was that the name resolution logic was (in the end)
calling #resolve_type_definition for the including type (the
definition context). In turn, this resolution is trying to
resolve the type again (since it is used for e.g. a field).
Since we don't have a definition for this type yet, it fails
there.

The fix is to make #resolve_context only resolve names, not
full types.