ruby: fix comparison of Typelib::Type with a class that is a direct subclass of Object
The "technique" used to so far to determine in Type.== whether we should
go on with the comparison fails if we were comparing Typelib::Type with
e.g.  Class.new. The reason was that we were looking at the superclass,
which is a bad idea in this case (the superclass in both cases is Object)