typemodel: change "s_available_containers" from pointer to object
fixes the test_suite so that it does not segfault (when no plugins can
be found) in test_model.cc:142 because "createContainer" is called,
which used the pointer without checking it.

the commit changes the pointer to be a static empty std::map, instead of
beeing sometimes created with new. removes null-pointer checks.

overhead? well, most of the time std::vectors are present and thus the
std::map would be created anyways... in case someone asks: on my 64bit
machine the size of the empty map is 48 bytes, versus 8 bytes for the
pointer.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>