test/ruby: remove not-needed-anymore requieres
a remnant from the "remove-dyncall-merge".

before, a "libtest_ruby" shared object using "test_rb_value.cc" was
compiled and loaded into some ruby-unittests. But, as far as I
understand this, it was only used by "test_value.rb".

I did somehow not check this... Might be because I never completely
deleted the build-folder, and hence the still-present shared object was
always found?

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>