typelib/exporter: remove deprecated function
the comment from 2008 says is deprecated. However, now that I get a
clear look at the signature and the body this function it looks actually
helpful in some cases... it will catch exceptions and return a boolean.
but its not consistent with the rest of the api.

anyways, is not used anywhere, hence removed... bäm.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>