TypelibBuilder: fix "anonymous" mess...
man, this whole stuff... two weeks old and already full of cruft ;-)

testing for "isAnonymousStructOrUnion" does not catch everything we
would like to catch. a quote from the Decl.h:

  /// isAnonymousStructOrUnion - Whether this is an anonymous struct
  /// or union. To be an anonymous struct or union, it must have been
  /// declared without a name and there must be no objects of this
  /// type declared, e.g.,
  /// @code
  ///   union { int i; float f; };
  /// @endcode
  /// is an anonymous union but neither of the following are:
  /// @code
  ///  union X { int i; float f; };
  ///  union { int i; float f; } obj;
  /// @endcode

so to catch the latter, we can test with "getIdentifier()" inside
"addRecord()"

remove the "string-compare-canonicalNames-to-contain-anonymous-hack"
which was bound to break, earlier than later... also added a fixme for a
similar hack -- testing for "sizeof"? hey, what can possibly go wrong...

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>