clang-importer: fix metadata entry "source_file_line" for Typelib::Type
before, the path to the header-file was stored into "orogen_include",
which is (as far as I see) is used by orogen at a later stage.  now,
"/path/to/file:line:column" is stored into source_file_line", as gccxml
does (gccxml.rb:445).

clang returns line-numbers starting with one, so the gccxml-related
trickery of adding 1 to the line-number is not needed.

renamed function and added comments for more claritry.

wondering: wouldn't it be a sleeker interface if
Typelib::Type::setPathToDefiningHeader() called
Typelib::Type::addMetaData() all by itself?

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>