typemodel: make "Type::ValidCategories" be a proper enum-value
and rename it to "NumberOfValidCategories"

"Typelib::Type::ValidCategories" was wrongly set to "Compound+1", a
remnant from 2005 -- impressive by itself ;-) the "Opaque" and
"Container" enums where then added in 2008 without adapting the value of
"ValidCategories". Seems like nobody is using this, so...

not sure: setting the start-value of the enum in "Typelib::Numeric"
to "Typelib::Type::ValidCategories" is nice but futile?

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>