clang-tlb-import: current set of tlb-files
as they are generated by the tool on my machine. note that the meta-data
contains the full path of the headers, which naturally contains my
home-folder...

the only example currently not working at all is "consts.h", the tlb is
excluded.

the others where not viewed at yet... commited just for later record.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>