clang-tlb-import: fill meta-data with "setHeaderPathForTypeFromDecl()"
before, for types coming from a macro-expansion, two locs where printed
by the SourceManager (and after that entered into the meta-data): the
locations of the macro and the expansion. now doing it manually just for
the expansion.

also now a pointer instead of a reference is given. makes the code
easier, because all other places use pointers to "Typelib::Type"
objects.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>