clang-tlb-import: renovate tests in "registerType()"
don't do string-manipulation to detect to-be-excluded types as it does
not make sense -- use the clang-API for this.

extended the test for "function pointer" to excludede "any pointer"
types. and removed the "sizeof" test... can find an example in the
test-suite where the missing test will create an error. in the contrary:
a (mean) test-case now passes where is should pass. updated tlb-file of
the test-case included.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>