clang-tlb-import: "addRecord()": move test for already added types to beginning
testing if the work will be futile before doing all the work might save
the work.

will remove superfluous add of a "numeric double" to "classes.tlb". was
created before because the "/classes/C1" have a double-field. but they
are also opaques and hence in the database from the beginning. no
information for these need to be known by typelib, the heavy-lifting
will be done by the user-supplied opaque-specific code.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>