typelib-clang-import: add new metadata function to fill "orogen_include"
kind-of experimental, not sure if usable by the orogen-stage...

will add the main-header if the type is defined there or earliest
non-main-header leading to the actual type. allows generating code more
safely.

affects all generated tlb-files because of the additional metadata entry
-- which is most of the time more or less the same as "source_file_line"

the interesting effect is the testcase containing eigen headers
(templates.tlb). now the correct first-level header "Eigen/Core" is
added to the database instead of some header deep inside the eigen-tree.

what is missing is the "pkg-config" bit "orogen_include"... this would
have to be done by a later massaging stage, as the clang-importer does
not know anything about pkg-config.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>