clang-tlb-import: re-enable "consts.hh", add "consts.tlb"
did crash before, works now...

"const" qualifiers have the problem that (at least with the
clang-importer) they are aliased in the typeli-registry to their
non-const version. thereby circumventing the whole type-annotation
system of C/C++... but as the type-information is only used to
copy/transport the data, this is not so bad. it does not make sense to
have a "const&" as datatype

altough it would be a cool feature: to generate the code somehow and let
it compile for in-process transports to allow language-based
zero-copy... whatever...

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>