clang-tlb-import: don't set the "orogen_include" metadata entry
while deactivating it: only return the _first_ header in the chain which
is a system-header, don't go any deeper into the rabbithole.

for the time beeing. in my view this is still a task for the
compiler-based tool, not for orogen. well, as the other codepath
currently works well skip this for later.

and update some of the tlb-files.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>