ruby: allow to select the C++ importer
Instead of completely replacing gccxml by clang, this commit
adds an impedance layer so that we can easily choose which
importer should be used.

This is controlled either programatically by setting Typelib::CXX.loader
to the underlying loader object (GCCXMLLoader or ClangLoader), or
from an environment variable (TYPELIB_CXX_LOADER) set to either
gccxml or clang. The default loader is still gccxml.