refactor memory layout handling
The MemoryLayout type was so far a typedef from std::vector. There
is the need to add more information than the plain layout (see
follow-up commits), so I chose to refactor it to make it a real class,
and used the opportunity to move most of the modification and
simplification code there instead of having it in the visitor.