fix: prevent multiple entries of the same name in opaque-registry
not sure if this is the right location to fix the problem of having the
same "opaques" in the "options[:opaques]" data-structure. but it
illustrates the problem: prevent the "type /base/Vector2d already
defined in registry" error from occuring.

this commit can surely be discussed. just my private hack to get it
working!

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>