exporter: add checks on export-filename, throw if bad
The string given as filename was not checked for validity after opening
before. This lead to silently _not_ writing to the given filename (if it
is an existing directory for example).

Now the function will throw() if the file could not be opened.
Additionally exceptions are activated for the ofstream-object so that
later failures (like a full disc or unmounted partition) will abort
the execution as well.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>