clang: DANGEROUS: ignore ABI namespaces that were introduced by gcc5
This commit ignores all namespaces like __1 (macos) and __cxx11
to make sure the generated typelib files can be used.
This commit is dangerous because it ignores the safety net which was
introduced to make sure a ABI compatibility between cxx11 and older
versions. But we have currently no way to represend this in the
typelib system.