fix TypeBuilder w.r.t. template-like constructs
The type builder name parsing would fail on e.g.
   /a/type</int[32]>

This is currently only used in the tlb importer, which once-upon-a-time
was able to load TLBs where e.g. arrays would not be defined. This is
not the case anymore (which is probably why we have not hit the problem
... yet)