integrate trailing-skip removal in simplify
There was a bug in removeTrailingSkips that was the sibling of
https://github.com/orocos-toolchain/typelib/issues/55. That is,
the removeTrailingSkips method would corrupt the layout if the
one-before-last element was 8 (FLAG_SKIP), regardless of the rest,
as e.g.

   FLAG_ARRAY 8
      FLAG_MEMCPY 8
   FLAG_END

This commit removes removeTrailingSkip, instead integrating the
functionality in simplify() -- which is working while parsing the
ops structure forward instead of backwards.