normalize the File name attributes by expanding them
castxml will report a relative path if given explicitely on
the command line, or if it is resolves through the include path
and the include path has .. in it.

That is, -I/some/path/../bla will lead to a file path that
includes ..

Expand all of them. There was already some normalization done
at the end of the chain, but this was failing if the relative
path is given on the command line (thus completely making the
load fail).