feat: standardize strings on /int8_t
A well known "feature" of C (and therefore C++) is that the signed-ness
of char is architecture-dependent. This is (very unfortunately)
represented as-is into the Typelib representation of C++ base types,
which leads to typelib thinking that strings generated e.g. ARM are
different than strings generated on a x86.

The "right" fix I've been thinking about for years now would be to stop
using an int8 type for strings, but add a completely separate character
type. However, this is an amount of work I'm not willing to do right
now.

So, in the meantime, standardize on int8 on all platforms.